ai-interviewer/pyproject.toml
[tool.poetry]
name = "ai-interviewer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<4.0,>=3.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
anthropic = "^0.39.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
